# grov-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grōv Restoration Platform API",
        "description": "Production-quality REST API backend for Grōv field restoration tracking, monitoring, community hub, and leaderboard.",
        "contact": {
            "name": "Grōv Engineering Team",
            "email": "support@grov.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Grōv Primary API Server v1"
        }
    ],
    "paths": {
...
        "/monitoring/records": {
            "get": {
                "tags": [
                    "Monitoring System"
                ],
                "summary": "List monitoring history records",
                "operationId": "5d6ed0dbbab3363f17964fce31abd19d",
                "responses": {
                    "200": {
                        "description": "Monitoring records retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Monitoring System"
                ],
                "summary": "Submit a monitoring record",
                "description": "Records observation count, survival status, notes, and photos for a past activity.",
                "operationId": "5f83763a73828ca532b4edd4719f788a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "activity_id",
                                    "observation_date",
                                    "observed_count",
                                    "established_count",
                                    "surviving_count",
                                    "dead_count",
                                    "condition"
                                ],
                                "properties": {
                                    "activity_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "observation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-27"
                                    },
                                    "observed_count": {
                                        "type": "integer",
                                        "example": 96
                                    },
                                    "established_count": {
                                        "type": "integer",
                                        "example": 82
                                    },
                                    "surviving_count": {
                                        "type": "integer",
                                        "example": 78
                                    },
                                    "dead_count": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "condition": {
                                        "type": "string",
                                        "enum": [
                                            "Good",
                                            "Fair",
                                            "Poor",
                                            "Unknown"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "photos[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Monitoring record submitted successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/activities/{activity_id}/monitoring": {
            "get": {
                "tags": [
                    "Monitoring System"
                ],
                "summary": "List monitoring observations for specific activity",
                "operationId": "844c61f329194c82e6dc4ad79dacd88f",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity monitoring records retrieved successfully"
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/species": {
            "get": {
                "tags": [
                    "Species Catalogue"
                ],
                "summary": "List species catalogue",
                "description": "Returns list of active plant and seed species with optional category and native filters.",
                "operationId": "fdb5a82dd855a934911ae708fadcc872",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "conifer",
                                "deciduous",
                                "mangrove",
                                "shrub",
                                "mixed"
                            ]
                        }
                    },
                    {
                        "name": "is_native",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Species list retrieved successfully"
                    }
                }
            }
        },
        "/species/{id}": {
            "get": {
                "tags": [
                    "Species Catalogue"
                ],
                "summary": "Get species details",
                "description": "Returns detailed information, suitable zones, description, and planting guidance for a single species.",
                "operationId": "a3083ebe764bf360ec4ca5aeb1938fd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Species details retrieved successfully"
                    },
                    "404": {
                        "description": "Species not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum bearer token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}